pyproject.toml
[project]
name = "course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopy>=2.4.1",
    "google-adk>=1.5.0",
    "jupyterlab>=4.4.4",
    "litellm>=1.74.0.post1",
    "numpy>=2.3.1",
    "openmeteo-requests>=1.5.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "timezonefinder>=6.5.9",
]
